+++ src/__internal__/fieldset/fieldset.spec.js
@@ -126,4 +126,16 @@ describe('Fieldset', () => {
       expect(icon.props()[validationType]).toEqual('Message');
     });
   });
+  it('add an asterisk after the text when the field is mandatory', () => {
+    assertStyleMatch(
+      {
+        content: "'*'",
+        color: '#C7384F',
+        fontWeight: '700',
+        marginLeft: '8px'
+      },
+      mount(<StyledLegendContainer isRequired />),
+      { modifier: '::after' }
+    );
+  });
 });