+++ src/__internal__/fieldset/fieldset.style.js
@@ -50,6 +50,17 @@ const StyledLegendContainer = styled.div`
     line-height: 24px;
   }
 
+  ${({
+    isRequired, theme
+  }) => isRequired && css`
+    ::after {
+      content: '*';
+      color: ${theme.colors.asterisk};
+      font-weight: 700;
+      margin-left: ${theme.spacing}px;
+    }
+  `}
+
   ${({ styleOverride }) => styleOverride};
 `;